+++ lib/mongoid_orderable.rb
@@ -25,7 +25,8 @@ require 'mongoid/orderable/generators/movable'
 require 'mongoid/orderable/generators/position'
 require 'mongoid/orderable/generators/scope'
 require 'mongoid/orderable/generators/helpers'
+require 'mongoid/orderable/handlers/base'
 require 'mongoid/orderable/handlers/document'
 require 'mongoid/orderable/handlers/embedded'
-require 'mongoid/orderable/handlers/transactional'
+require 'mongoid/orderable/handlers/transaction'
 require 'mongoid/orderable/installer'
